v2
Prompt 1:
Let's take your code:
[Insert code]
and edit it. It currently finds None as the CSV and JSON URLs, and it leaves the HTML URL as empty. Here are the buttons that need to be clicked to ascertain the real URLs.
[Show the buttons]

Prompt 2:
Keep in mind that, once you get the HTML URL, it changes the page. You have to go back to the previous page after clicking the export_btn in order to click on the csv_btn and json_btn to get their respective URLs